Bumping version
We are adding a new model, so we bump it to version 0.10.0.
After a co-programming session, it seems that the model is
behaving as expected. The main thing is that the mid-risk level
has low prevalence in the population, so it is not playing an important
role in the process