Adding mixing risk model (#101)
* Adding mixing risk model

* Working on the tests

* Adding details to the contacttracing class

* Refactoring testing

* Bumping version

We are adding a new model, so we bumped it to version 0.10.0.
After a pair-programming session, it seems that the model is
behaving as expected. The main thing is that the mid-risk level
has low prevalence in the population, so it is not playing an important
role in the process